requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-core==2.14.0
orjson==3.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.1.4
openpyxl==3.1.2
librosa==0.10.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
faster-whisper==0.9.0
pydub==0.25.1
yt-dlp==2023.11.16
beautifulsoup4==4.12.2
requests==2.31.0
aiofiles==23.2.1